Straighten the class hierarchy in edb.schema
This removes `InheritingObjectBase` and `DerivableObjectBase` in favor
of just `InheritingObject` and `DerivableObject`.
