+++ mypy.ini
@@ -397,22 +397,6 @@ warn_unused_ignores = True
 warn_return_any = True
 no_implicit_reexport = True
 
-[mypy-edb.schema.derivable]
-follow_imports = True
-ignore_errors = False
-# Equivalent of --strict on the command line:
-disallow_subclassing_any = True
-disallow_any_generics = True
-# disallow_untyped_calls = True
-disallow_untyped_defs = True
-disallow_incomplete_defs = True
-check_untyped_defs = True
-disallow_untyped_decorators = True
-no_implicit_optional = True
-warn_unused_ignores = True
-warn_return_any = True
-no_implicit_reexport = True
-
 [mypy-edb.schema.inheriting]
 follow_imports = True
 ignore_errors = False
